moved modmail methods to different file. same with nsfw monitoring.
