+++ .gitignore
@@ -4,5 +4,4 @@ settings.py
 settings.pyc
 
 startModeratelyHelpfulBot.sh
-/ModeratelyHelpfulBot.py
 praw.ini
