+++ workingdata.py
@@ -4,7 +4,7 @@ class WorkingData:
     to_update_list = True
     ri = None
     sub_dict = {}
-    nsfw_monitoring_subs = []
+    nsfw_monitoring_subs = {}
 
     def __init__(self):
